reworked cofiguration, cleanup, new VId&PId, repo cleanup
New configurable parameters have been moved to a struct (EXPCFGData) for easier communication and to get around some of the fixed position status words.
Added an option to make CoilConfigs persistant, so they don't have to be activated every reboot.

Unused variables have been removed (mainly in MidiController.c).

Added *.d to gitignore and ran clean to get rid of compilation artifacts in the repo.

Changed the Vendor ID and Product IDs to use the ones we received from what is left of OpenMoko. IMPORTANT: The current version of the Stickfigurator does NOT know those Pids and will not find you stick if you decide to run form compiled code. The new version will be released this weekend